pyproject.toml
[tool.poetry]
name = "skuy-work"
version = "0.1.0"
description = ""
authors = ["Yosef <yosef@skuy.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-fasthtml = "^0.6.8"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
